Added many fixes to auth module.
* All forms in auth module are now passed the request object. This is
helpful for example when the Form wants to present the reCaptch field
after the user fails validation for three times and it wants to store
that attempt count in request.session.
* Now many of the view methods accept personal extra parameters. e.g.
login now accepts login_form_class (as well as form_class) to allow
user to specify their custom LoginForm. So the usual format is
<view_name>_<param_name>. This was needed since if user passes
form_class while using the include() then that would set the new
form_class as the Form for many methods including timezone_change(),
change_password() etc.
* Fixed few view methods which did not expect to receive unexpected
parameters, like confirm_email(). All the view methods might get
unexpected params if the user passes extra params while using include().
* Changed the Forms to remove the passed request param since the
original implementation does not use that and that is intended to be
used by subclasses.
* Modified ChangePasswordForm to not show the old password filed if the
current password is not set. If the user signs up via opened then
password is set to unusable (it is set to ! character which hashes
cannot generate). Making it impossible to change the password.